delete:delete the signup.go file as its code is refactored in the auth_handler.go
